+++ BillyCDK.DataAccess/Database/DBCollection.cs
@@ -1,0 +1,20 @@
+using BillyCDK.DataAccess.Models;
+using MongoDB.Driver;
+
+namespace BillyCDK.DataAccess.Database;
+
+public class DBCollection : IDBCollection
+{
+    private readonly IDBContext context;
+    private readonly IMongoDatabase dbInstance;
+
+    public DBCollection(IDBContext context)
+    {
+        this.context = context;
+        dbInstance = context.DBInstance;
+    }
+
+    public IMongoCollection<Player> Players => dbInstance.GetCollection<Player>("players");
+    public IMongoCollection<Game> Games => dbInstance.GetCollection<Game>("games");
+    public IMongoCollection<CDKey> CDKeys => dbInstance.GetCollection<CDKey>("cdkeys");
+}
