+++ BillyCDK.DataAccess/Models/Graphs/Query.graphql
