+++ BillyCDK.DataAccess/Models/InputCDK.cs
@@ -1,4 +1,4 @@
-﻿namespace BillyCDK.App.Models;
+﻿namespace BillyCDK.DataAccess.Models;
 
 public record InputCDKey(
     string Value,
