+++ BillyCDK.DataAccess/Models/InputGame.cs
@@ -1,4 +1,4 @@
-namespace BillyCDK.App.Models;
+namespace BillyCDK.DataAccess.Models;
 
 public record InputGame(
     string? DBName,
