+++ BillyCDK.Test/ObjectTests/PlayerTest.cs
@@ -1,7 +1,7 @@
 using System.Text.Json;
-using BillyCDK.App.Models;
-using BillyCDK.App.Services;
-using BillyCDK.App.Utilities;
+using BillyCDK.DataAccess.Models;
+using BillyCDK.DataAccess.Services;
+using BillyCDK.DataAccess.Utilities;
 using BillyCDK.Test.Infrastructure;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
